Separated sorting and creator functionality
