+++ src/app/chat/chat.component.html
@@ -23,7 +23,7 @@
 <div id="writemessage">
     <form name="message" action="" >
         <input name="usermsg" type="text" id="usermsg" size="120"  placeholder="Write your message..." />
-        <button (click)="createNewMessage()" id="submitmsg"><i class="fa fa-paper-plane" aria-hidden="true"  style="font-size: 25px; "></i></button>
+        <button (click)="callFilesCreatorMessage()" id="submitmsg"><i class="fa fa-paper-plane" aria-hidden="true"  style="font-size: 25px; "></i></button>
     </form>
 </div>
